move atomic intrinsics to their own file
