+++ rust-version
@@ -1,1 +1,1 @@
-db41351753df840773ca628d8daa040e95d00eef
+880416180b0a9ee1141c07d4d17667edb77daebd
